+++ repl/CHANGELOG.md
@@ -11,7 +11,8 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 * Add support for evaluation of lambda terms in environment
 * Add command to expand terms only (no beta-reduction)
 * Add command to apply beta-reduction (no expansion of predefined terms)
-* Add `let` command to add new bindings to the environment or replace existing ones
+* Add `:let` command to add new bindings to the environment or replace existing ones
+* Add commands to maintain bindings in the environment (`:ls-env`, `clr-env`, `ld-env`)
 * Improve help text
 * Upgrade `rustyline` to version 2.0